third_party/WebKit/LayoutTests/fast/events/platform-wheelevent-paging-y-in-scrolling-page.html

<html>
    <head>
        <script src="../../resources/js-test.js"></script>
        <script>
            window.jsTestIsAsync = true;
            var givenScrollTop = 2; // Only page is scrolled per event, regardless of this number.
            var givenScrollLeft = 0;
            var expectedScrollTop = 525; // Window is 800x600. Scroll 87.5% of visible.
            var expectedScrollLeft = 0;
            var event;
            var div;

            function runTest()
            {
                document.body.addEventListener("mousewheel", mousewheelHandler, false);

                if (window.eventSender) {
                    eventSender.mouseMoveTo(100, 110);
                    eventSender.continuousMouseScrollBy(-window.givenScrollLeft, -window.givenScrollTop, true);
                }

                // This test expects failure on mac: crbug.com/552694.
                shouldBecomeEqual("document.scrollingElement.scrollTop == window.expectedScrollTop && " +
                    "document.scrollingElement.scrollLeft == window.expectedScrollLeft", "true", finishJSTest);
            }

            function mousewheelHandler(e)
            {
                event = e;
                shouldBe("event.wheelDeltaY", "window.givenScrollTop * -3");
                shouldBe("event.wheelDeltaX", "0");

                if (e.wheelDeltaY)
                    shouldBe("event.wheelDelta", "window.givenScrollTop * -3");
                else
                    shouldBe("event.wheelDelta", "0");
            }
        </script>
    </head>

    <body style="margin:0" onload="runTest()">
        <div style="height:900px;width:400px">
          <div style="background-color:red;height:900px;width:200px;position:relative;left:0px;top:0px"></div>
          <div style="background-color:green;height:900px;width:200px;position:relative;left:200px;top:-900px"></div>
        </div>
        <div style="height:900px;width:400px">
          <div style="background-color:blue;height:900px;width:200px;position:relative;left:0px;top:0px"></div>
          <div style="background-color:yellow;height:900px;width:200px;position:relative;left:200px;top:-900px"></div>
        </div>
        <div id="console"></div>
    </body>
</html>
